Python formatting fixes
