+++ server/controllers/api/show/script/config.py
@@ -1,0 +1,26 @@
+from typing import List
+
+from models.session import Session
+from utils.web.base_controller import BaseAPIController
+from utils.web.route import ApiRoute, ApiVersion
+
+
+@ApiRoute("show/script/config", ApiVersion.V1)
+class ScriptStatusController(BaseAPIController):
+    def get(self):
+        with self.make_session() as session:
+            editors: List[Session] = (
+                session.query(Session).filter(Session.is_editor).all()
+            )
+            if editors:
+                current_editor = editors[0].internal_id
+            else:
+                current_editor = None
+
+            data = {
+                "canRequestEdit": len(editors) == 0,
+                "currentEditor": current_editor,
+            }
+
+            self.set_status(200)
+            self.finish(data)
