+++ server/controllers/api/show/script/script.py
@@ -4,7 +4,6 @@ from typing import List, Optional
 from sqlalchemy import func
 from tornado import escape
 
-from models.cue import CueAssociation
 from models.script import (
     Script,
     ScriptCuts,
@@ -12,329 +11,13 @@ from models.script import (
     ScriptLinePart,
     ScriptLineRevisionAssociation,
     ScriptRevision,
-    StageDirectionStyle,
 )
-from models.session import Session
 from models.show import Show
 from rbac.role import Role
-from schemas.schemas import (
-    ScriptLineSchema,
-    ScriptRevisionsSchema,
-    StageDirectionStyleSchema,
-)
+from schemas.schemas import ScriptLineSchema
 from utils.web.base_controller import BaseAPIController
 from utils.web.route import ApiRoute, ApiVersion
 from utils.web.web_decorators import no_live_session, requires_show
-
-
-@ApiRoute("show/script/config", ApiVersion.V1)
-class ScriptStatusController(BaseAPIController):
-    def get(self):
-        with self.make_session() as session:
-            editors: List[Session] = (
-                session.query(Session).filter(Session.is_editor).all()
-            )
-            if editors:
-                current_editor = editors[0].internal_id
-            else:
-                current_editor = None
-
-            data = {
-                "canRequestEdit": len(editors) == 0,
-                "currentEditor": current_editor,
-            }
-
-            self.set_status(200)
-            self.finish(data)
-
-
-@ApiRoute("show/script/revisions", ApiVersion.V1)
-class ScriptRevisionsController(BaseAPIController):
-
-    @requires_show
-    def get(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-        revisions_schema = ScriptRevisionsSchema()
-
-        with self.make_session() as session:
-            show: Show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-
-                if script:
-                    revisions = [revisions_schema.dump(c) for c in script.revisions]
-                    self.set_status(200)
-                    self.finish(
-                        {
-                            "current_revision": script.current_revision,
-                            "revisions": revisions,
-                        }
-                    )
-                else:
-                    self.set_status(404)
-                    self.finish({"message": "404 script not found"})
-            else:
-                self.set_status(404)
-                self.finish({"message": "404 show not found"})
-
-    @requires_show
-    @no_live_session
-    async def post(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show = session.query(Show).get(show_id)
-            if show:
-                data = escape.json_decode(self.request.body)
-
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                if not script:
-                    self.set_status(404)
-                    await self.finish({"message": "404 script not found"})
-                    return
-                self.requires_role(script, Role.WRITE)
-
-                current_rev_id = script.current_revision
-                if not current_rev_id:
-                    self.set_status(404)
-                    await self.finish({"message": "404 script revision not found"})
-                    return
-
-                current_rev: ScriptRevision = session.query(ScriptRevision).get(
-                    current_rev_id
-                )
-                if not current_rev:
-                    self.set_status(404)
-                    await self.finish({"message": "404 script revision not found"})
-                    return
-
-                max_rev = (
-                    session.query(func.max(ScriptRevision.revision))
-                    .filter(ScriptRevision.script_id == script.id)
-                    .one()[0]
-                )
-
-                description: str = data.get("description", None)
-                if not description:
-                    self.set_status(400)
-                    await self.finish({"message": "Description missing"})
-                    return
-
-                now_time = datetime.utcnow()
-                new_rev = ScriptRevision(
-                    script_id=script.id,
-                    revision=max_rev + 1,
-                    created_at=now_time,
-                    edited_at=now_time,
-                    description=description,
-                    previous_revision_id=current_rev.id,
-                )
-                session.add(new_rev)
-                session.flush()
-                for line_association in current_rev.line_associations:
-                    new_rev.line_associations.append(
-                        ScriptLineRevisionAssociation(
-                            revision_id=new_rev.id,
-                            line_id=line_association.line_id,
-                            next_line_id=line_association.next_line_id,
-                            previous_line_id=line_association.previous_line_id,
-                        )
-                    )
-                for cue_association in current_rev.cue_associations:
-                    new_rev.cue_associations.append(
-                        CueAssociation(
-                            revision_id=new_rev.id,
-                            line_id=cue_association.line_id,
-                            cue_id=cue_association.cue_id,
-                        )
-                    )
-                for cut_association in current_rev.line_part_cuts:
-                    new_rev.line_part_cuts.append(
-                        ScriptCuts(
-                            revision_id=new_rev.id,
-                            line_part_id=cut_association.line_part_id,
-                        )
-                    )
-
-                script.current_revision = new_rev.id
-                session.commit()
-
-                self.set_status(200)
-                await self.finish(
-                    {"id": new_rev.id, "message": "Successfully added script revision"}
-                )
-                await self.application.ws_send_to_all(
-                    "NOOP", "GET_SCRIPT_REVISIONS", {}
-                )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
-
-    @requires_show
-    @no_live_session
-    async def delete(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show: Show = session.query(Show).get(show_id)
-            if show:
-                data = escape.json_decode(self.request.body)
-
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                if not script:
-                    self.set_status(404)
-                    await self.finish({"message": "404 script not found"})
-                    return
-                self.requires_role(script, Role.WRITE)
-
-                rev_id: int = data.get("rev_id", None)
-                if not rev_id:
-                    self.set_status(400)
-                    await self.finish({"message": "Revision missing"})
-                    return
-
-                rev: ScriptRevision = session.query(ScriptRevision).get(rev_id)
-                if not rev:
-                    self.set_status(404)
-                    await self.finish({"message": "Revision not found"})
-                    return
-
-                if rev.script_id != script.id:
-                    self.set_status(400)
-                    await self.finish(
-                        {"message": "Revision is not for the current script"}
-                    )
-                    return
-
-                if rev.revision == 1:
-                    self.set_status(400)
-                    await self.finish(
-                        {"message": "Cannot delete first script revision"}
-                    )
-                    return
-
-                changed_rev = False
-                if script.current_revision == rev.id:
-                    changed_rev = True
-                    if rev.previous_revision_id:
-                        script.current_revision = rev.previous_revision_id
-                    else:
-                        first_rev: ScriptRevision = (
-                            session.query(ScriptRevision)
-                            .filter(
-                                ScriptRevision.script_id == script.id,
-                                ScriptRevision.revision == 1,
-                            )
-                            .one()
-                        )
-                        script.current_revision = first_rev.id
-
-                session.delete(rev)
-                session.commit()
-
-                self.set_status(200)
-                await self.finish({"message": "Successfully deleted script revision"})
-                if changed_rev:
-                    await self.application.ws_send_to_all(
-                        "NOOP", "SCRIPT_REVISION_CHANGED", {}
-                    )
-                else:
-                    await self.application.ws_send_to_all(
-                        "NOOP", "GET_SCRIPT_REVISIONS", {}
-                    )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
-
-
-@ApiRoute("show/script/revisions/current", ApiVersion.V1)
-class ScriptCurrentRevisionController(BaseAPIController):
-
-    @requires_show
-    def get(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show: Show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-
-                if script:
-                    self.set_status(200)
-                    self.finish(
-                        {
-                            "current_revision": script.current_revision,
-                        }
-                    )
-                else:
-                    self.set_status(404)
-                    self.finish({"message": "404 script not found"})
-            else:
-                self.set_status(404)
-                self.finish({"message": "404 show not found"})
-
-    @requires_show
-    @no_live_session
-    async def post(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show = session.query(Show).get(show_id)
-            if show:
-                self.requires_role(show, Role.WRITE)
-                data = escape.json_decode(self.request.body)
-
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                if not script:
-                    self.set_status(404)
-                    await self.finish({"message": "404 script not found"})
-                    return
-
-                new_rev_id: int = data.get("new_rev_id", None)
-                if not new_rev_id:
-                    self.set_status(400)
-                    await self.finish({"message": "New revision missing"})
-                    return
-
-                new_rev: ScriptRevision = session.query(ScriptRevision).get(new_rev_id)
-                if not new_rev:
-                    self.set_status(404)
-                    await self.finish({"message": "New revision not found"})
-                    return
-
-                if new_rev.script_id != script.id:
-                    self.set_status(400)
-                    await self.finish(
-                        {"message": "New revision is not for the current script"}
-                    )
-                    return
-
-                script.current_revision = new_rev.id
-                session.commit()
-
-                self.set_status(200)
-                await self.finish({"message": "Successfully changed script revision"})
-                await self.application.ws_send_to_all(
-                    "NOOP", "SCRIPT_REVISION_CHANGED", {}
-                )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
 
 
 @ApiRoute("/show/script", ApiVersion.V1)
@@ -1102,232 +785,3 @@ class ScriptMaxPageController(BaseAPIController):
                 self.set_status(404)
                 self.finish({"message": "404 show not found"})
                 return
-
-
-@ApiRoute("/show/script/stage_direction_styles", ApiVersion.V1)
-class StageDirectionStylesController(BaseAPIController):
-    @requires_show
-    def get(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        stage_direction_style_schema = StageDirectionStyleSchema()
-        with self.make_session() as session:
-            show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                stage_direction_styles = [
-                    stage_direction_style_schema.dump(style)
-                    for style in script.stage_direction_styles
-                ]
-
-                self.set_status(200)
-                self.finish({"styles": stage_direction_styles})
-            else:
-                self.set_status(404)
-                self.finish({"message": "404 show not found"})
-                return
-
-    @requires_show
-    @no_live_session
-    async def post(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                self.requires_role(script, Role.WRITE)
-                data = escape.json_decode(self.request.body)
-
-                description: str = data.get("description", None)
-                if not description:
-                    self.set_status(400)
-                    await self.finish({"message": "Description missing"})
-                    return
-
-                bold: bool = data.get("bold", False)
-                italic: bool = data.get("italic", False)
-                underline: bool = data.get("underline", False)
-
-                text_format: str = data.get("textFormat", None)
-                if not text_format or text_format not in ["default", "upper", "lower"]:
-                    self.set_status(400)
-                    await self.finish({"message": "Text format missing or invalid"})
-                    return
-
-                text_colour: str = data.get("textColour", None)
-                if not text_colour:
-                    self.set_status(400)
-                    await self.finish({"message": "Text colour missing"})
-                    return
-
-                enable_background_colour: bool = data.get(
-                    "enableBackgroundColour", False
-                )
-                background_colour: str = data.get("backgroundColour", None)
-                if enable_background_colour and not background_colour:
-                    self.set_status(400)
-                    await self.finish({"message": "Background colour missing"})
-                    return
-
-                new_style = StageDirectionStyle(
-                    script_id=script.id,
-                    description=description,
-                    bold=bold,
-                    italic=italic,
-                    underline=underline,
-                    text_format=text_format,
-                    text_colour=text_colour,
-                    enable_background_colour=enable_background_colour,
-                    background_colour=background_colour,
-                )
-                session.add(new_style)
-                session.commit()
-
-                self.set_status(200)
-                await self.finish(
-                    {
-                        "id": new_style.id,
-                        "message": "Successfully added stage direction style",
-                    }
-                )
-
-                await self.application.ws_send_to_all(
-                    "NOOP", "GET_STAGE_DIRECTION_STYLES", {}
-                )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
-
-    @requires_show
-    @no_live_session
-    async def patch(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show: Show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                self.requires_role(script, Role.WRITE)
-                data = escape.json_decode(self.request.body)
-
-                style_id = data.get("id", None)
-                if not style_id:
-                    self.set_status(400)
-                    await self.finish({"message": "ID missing"})
-                    return
-
-                style: StageDirectionStyle = session.query(StageDirectionStyle).get(
-                    style_id
-                )
-                if not style:
-                    self.set_status(404)
-                    await self.finish(
-                        {"message": "404 stage direction style not found"}
-                    )
-                    return
-
-                description: str = data.get("description", None)
-                if not description:
-                    self.set_status(400)
-                    await self.finish({"message": "Description missing"})
-                    return
-
-                bold: bool = data.get("bold", False)
-                italic: bool = data.get("italic", False)
-                underline: bool = data.get("underline", False)
-
-                text_format: str = data.get("textFormat", None)
-                if not text_format or text_format not in ["default", "upper", "lower"]:
-                    self.set_status(400)
-                    await self.finish({"message": "Text format missing or invalid"})
-                    return
-
-                text_colour: str = data.get("textColour", None)
-                if not text_colour:
-                    self.set_status(400)
-                    await self.finish({"message": "Text colour missing"})
-                    return
-
-                enable_background_colour: bool = data.get(
-                    "enableBackgroundColour", False
-                )
-                background_colour: str = data.get("backgroundColour", None)
-                if enable_background_colour and not background_colour:
-                    self.set_status(400)
-                    await self.finish({"message": "Background colour missing"})
-                    return
-
-                style.description = description
-                style.bold = bold
-                style.italic = italic
-                style.underline = underline
-                style.text_format = text_format
-                style.text_colour = text_colour
-                style.enable_background_colour = enable_background_colour
-                style.background_colour = background_colour
-                session.commit()
-
-                self.set_status(200)
-                await self.finish(
-                    {"message": "Successfully edited stage direction style"}
-                )
-
-                await self.application.ws_send_to_all(
-                    "NOOP", "GET_STAGE_DIRECTION_STYLES", {}
-                )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
-
-    @requires_show
-    @no_live_session
-    async def delete(self):
-        current_show = self.get_current_show()
-        show_id = current_show["id"]
-
-        with self.make_session() as session:
-            show: Show = session.query(Show).get(show_id)
-            if show:
-                script: Script = (
-                    session.query(Script).filter(Script.show_id == show.id).first()
-                )
-                self.requires_role(script, Role.WRITE)
-                data = escape.json_decode(self.request.body)
-
-                style_id = data.get("id", None)
-                if not style_id:
-                    self.set_status(400)
-                    await self.finish({"message": "ID missing"})
-                    return
-
-                entry: StageDirectionStyle = session.get(StageDirectionStyle, style_id)
-                if entry:
-                    session.delete(entry)
-                    session.commit()
-
-                    self.set_status(200)
-                    await self.finish(
-                        {"message": "Successfully deleted stage direction style"}
-                    )
-
-                    await self.application.ws_send_to_all(
-                        "NOOP", "GET_STAGE_DIRECTION_STYLES", {}
-                    )
-                else:
-                    self.set_status(404)
-                    await self.finish(
-                        {"message": "404 stage direction style not found"}
-                    )
-            else:
-                self.set_status(404)
-                await self.finish({"message": "404 show not found"})
