+++ server/digi_server/logger.py
@@ -62,9 +62,8 @@ def add_logging_level(level_name, level_num, method_name=None):
 
     def log_for_level(self, message, *args, **kwargs):
         if self.isEnabledFor(level_num):
-            self._log(
-                level_num, message, args, **kwargs
-            )  # pylint: disable=protected-access
+            # pylint: disable=protected-access
+            self._log(level_num, message, args, **kwargs)
 
     def log_to_root(message, *args, **kwargs):
         logging.log(level_num, message, *args, **kwargs)
